Solution to Sort Students by Marks

+++ Sort Students by Marks@@67877d61dc107486ad5c2748/code.c
@@ -7,7 +7,7 @@ int main() {
     for(int i=0;i<n;i++){
         scanf("%d %s %f",&roll_num[i],name[i],&marks[i]);
     }
-    int max_marks=marks[0];
+    float max_marks=marks[0];
     for(int i=1;i<n;i++){
         if(marks[i]>max_marks){
             max_marks=marks[i];
